Move DNS takeover from pentest to discover
Takeover detection checks whether a record can be claimed; it never
claims it. The CNAME lookup and the HTTP request both go to the
third-party provider the record dangles at, never to infrastructure the
target still controls. Nothing about it is a pentest, and the ontology
tool already declares Family.RECONNAISSANCE — only the CLI path said
otherwise.

Renames the stage-named Fern types to match, and the package to match
its new siblings (cctld, subdomain). The genuine pentest action here
would be claiming the dangling resource to prove the finding; that is
state-changing, belongs to a later phase, and does not exist.

Takeover was the only command under `pentest`, so the stage goes with
it and osintscan is now discover-only — nothing left in it interrogates
a target's own service.
