--- cmd/pentest.go
@@ -1,134 +1,0 @@
-package cmd
-
-import (
-	"errors"
-
-	dnsfern "github.com/Method-Security/osintscan/generated/go/pentest/dns"
-	takeover "github.com/Method-Security/osintscan/internal/pentest/dns/takeover"
-	"github.com/Method-Security/osintscan/utils"
-	"github.com/spf13/cobra"
-)
-
-func (a *OsintScan) InitPentestCommand() {
-	pentestCmd := &cobra.Command{
-		Use:   "pentest",
-		Short: "Run penetration testing modules",
-		Long:  `Run penetration testing modules to identify vulnerabilities in DNS and related infrastructure.`,
-	}
-
-	pentestDNSCmd := &cobra.Command{
-		Use:   "dns",
-		Short: "Run DNS-related penetration tests",
-		Long:  `Run penetration testing modules focused on DNS, such as subdomain takeover detection and DNS misconfiguration checks.`,
-	}
-
-	pentestCmd.AddCommand(pentestDNSCmd)
-
-	pentestDNSTakeoverCmd := &cobra.Command{
-		Use:   "takeover",
-		Short: "Detect potential subdomain takeovers",
-		Long:  `Analyze the provided targets to identify DNS records that may be vulnerable to subdomain takeover attacks, using known fingerprints and heuristics.`,
-		Run: func(cmd *cobra.Command, args []string) {
-			// Get targets
-			targets, err := cmd.Flags().GetStringSlice("targets")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			filePaths, err := cmd.Flags().GetStringSlice("target-files")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			fileTargets, err := utils.GetEntriesFromFiles(filePaths)
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-
-			allTargets := append(targets, fileTargets...)
-			if len(allTargets) == 0 {
-				a.OutputSignal.AddError(errors.New("no targets specified"))
-				return
-			}
-
-			// Config
-			fingerprintsPath, err := cmd.Flags().GetString("fingerprints-file")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			fingerprints, err := takeover.RetrieveFingerprints(fingerprintsPath)
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			if len(fingerprints) == 0 {
-				a.OutputSignal.AddError(errors.New("no fingerprints found"))
-				return
-			}
-
-			successfulOnly, err := cmd.Flags().GetBool("successful-only")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-
-			verifyTLS, err := cmd.Flags().GetBool("verify-tls")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-
-			timeout, err := cmd.Flags().GetInt("timeout")
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-
-			config := getDNSTakeoverConfig(allTargets, fileTargets, fingerprintsPath, successfulOnly, verifyTLS, timeout)
-
-			report, err := takeover.DetectDomainTakeover(cmd.Context(), config, fingerprints)
-			if err != nil {
-				a.OutputSignal.AddError(err)
-				return
-			}
-			a.OutputSignal.Content = report
-		},
-	}
-
-	// Target Flags
-	pentestDNSTakeoverCmd.Flags().StringSlice("targets", []string{}, "A list of URLs or domains to analyze for takeover vulnerabilities")
-	pentestDNSTakeoverCmd.Flags().StringSlice("target-files", []string{}, "File paths containing lists of targets to analyze for takeover vulnerabilities")
-	// Config Flags
-	pentestDNSTakeoverCmd.Flags().String("fingerprints-file", "", "Path to the JSON file containing service fingerprints for takeover detection")
-	pentestDNSTakeoverCmd.Flags().Bool("successful-only", false, "Show only confirmed successful takeovers in the results")
-	pentestDNSTakeoverCmd.Flags().Bool("verify-tls", false, "Verify TLS certificates when making HTTPS requests during takeover analysis")
-	pentestDNSTakeoverCmd.Flags().Int("timeout", 180, "Timeout in seconds for each takeover check request")
-
-	// Mark Required Flags
-	_ = pentestDNSTakeoverCmd.MarkFlagRequired("targets")
-	pentestDNSCmd.AddCommand(pentestDNSTakeoverCmd)
-
-	a.RootCmd.AddCommand(pentestCmd)
-}
-
-func getDNSTakeoverConfig(targets []string, targetFiles []string, fingerprintsFile string, successfulOnly bool, verifyTLS bool, timeout int) dnsfern.PentestDomainTakeoverConfig {
-	config := dnsfern.PentestDomainTakeoverConfig{
-		Targets:        targets,
-		TargetFiles:    targetFiles,
-		SuccessfulOnly: successfulOnly,
-		VerifyTls:      verifyTLS,
-	}
-	if fingerprintsFile != "" {
-		config.FingerprintsFile = &fingerprintsFile
-	}
-
-	if timeout >= 0 {
-		config.Timeout = timeout
-	} else {
-		config.Timeout = 30
-	}
-
-	return config
-}
