+++ mkdocs.yml
@@ -15,7 +15,6 @@ nav:
       - Overview: docs/index.md
       - Capabilities:
         - Discover: docs/discover.md
-        - Pentest: docs/pentest.md
   - Contributing:
       - How to contribute: community/community.md
       - Development:
